requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
pip==20.2.3
setuptools==50.3.0
typing-extensions==3.7.4.3
yarl==1.6.0